// src/generated/models/blood-group.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records - Appointment scheduling and management - Triage, vitals, and examination - Prescription and investigation management - IPD/OPD workflow 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const BloodGroup = {
    A: 'A+',
    A2: 'A-',
    B: 'B+',
    B2: 'B-',
    O: 'O+',
    O2: 'O-',
    Ab: 'AB+',
    Ab2: 'AB-'
} as const;

export type BloodGroup = typeof BloodGroup[keyof typeof BloodGroup];